+++ EmpyrionScripting/CsHelper/ItemAccessHelpers.cs
@@ -14,6 +14,6 @@ namespace EmpyrionScripting.CsHelper
         public int ConfigId(string name) => EmpyrionScripting.ConfigEcfAccess.FlatConfigBlockByName.TryGetValue(name, out var config) ? (int)config.Attr?.FirstOrDefault(A => A.Name == "Id")?.Value : 0;
         public EcfBlock ConfigById(int id) => EmpyrionScripting.ConfigEcfAccess.FlatConfigBlockById.TryGetValue(id, out var block) ? block : null;
         public EcfBlock ConfigByName(string name) => EmpyrionScripting.ConfigEcfAccess.FlatConfigBlockByName.TryGetValue(name, out var block) ? block : null;
-        public Dictionary<int, int> ResourcesForBlockById(int id) => EmpyrionScripting.ConfigEcfAccess.ResourcesForBlockById.TryGetValue(id, out var recipe) ? recipe : null;
+        public Dictionary<int, double> ResourcesForBlockById(int id) => EmpyrionScripting.ConfigEcfAccess.ResourcesForBlockById.TryGetValue(id, out var recipe) ? recipe : null;
     }
 }
